+++ training/src/anemoi/training/data/dataset/__init__.py
@@ -1,0 +1,13 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+from .ensdataset import EnsNativeGridDataset
+from .singledataset import NativeGridDataset
+
+__all__ = ["EnsNativeGridDataset", "NativeGridDataset"]
